Conflicts:
	build_scripts/build_usd.py

+++ pxr/usd/lib/sdf/childrenUtils.h
@@ -105,7 +105,7 @@ public:
         const SdfLayerHandle &layer,
         const SdfPath &path,
         const typename ChildPolicy::ValueType& value,
-        size_t index);
+        int index);
 
     /// Remove the child identified by \a key.
     SDF_API
@@ -127,7 +127,7 @@ public:
         const SdfPath &path,
         const typename ChildPolicy::ValueType& value,
         const typename ChildPolicy::FieldType& newName,
-        size_t index);
+        int index);
 
     /// Remove the child identified by \a key.
     SDF_API
@@ -148,7 +148,7 @@ public:
         const SdfPath &path,
         const typename ChildPolicy::ValueType& value,
         const typename ChildPolicy::FieldType& newName,
-        size_t index,
+        int index,
         std::string* whyNot);
 
     /// Returns \c true if the child of \p path identified by \p key can
